requirements.txt
modoboa>=1.7.2
django-phonenumber-field==1.2.0
django-webpack-loader
